requirements.txt
alembic==1.14.1
sqlalchemy==1.4.31
fastapi==0.73.0
asyncpg==0.30.0
aiofiles==23.2.1
bcrypt==4.2.1
uvicorn==0.17.4
python-dotenv==0.19.2
python-multipart==0.0.17
aiohttp==3.10.6